Changed some Model datatypes for account. Previously, Account was creating a relationship between the static tables, but now it just stores a byte/TINYINT for the relationship between them.

+++ GnuPlusLinuxDAL/Account/Class.cs
@@ -5,7 +5,7 @@ namespace GnuPlusLinuxDAL
     public class Class
     {
 		[Key]
-		public int ClassId { get; set; }
+		public byte ClassId { get; set; }
 
 		[Required]
 		public string ClassName { get; set; }
